### 2023-05-31 #5506 @editorialbot
[PRE REVIEW]: OasisMove: A Verified and Validated FEniCS-based Computational Fluid Dynamics Solver for Moving Domains
**Submitting author:** @hkjeldsberg (<a href="http://orcid.org/0000-0002-7764-4248">Henrik Aasen Kjeldsberg</a>)
**Repository:** https://github.com/KVSlab/OasisMove
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.2.0
