+++ paper/paper.bib
@@ -256,12 +256,13 @@
 }
 
 @article{kjeldsberg2023vampy,
+  doi={10.21105/joss.05278},
   title={VaMPy: An Automated and Objective Pipeline for Modeling Vascular Geometries},
   author={Kjeldsberg, Henrik A and Bergersen, Aslak W and Valen-Sendstad, Kristian},
   journal={Journal of Open Source Software},
-  volume={4},
-  number={35},
-  pages={1065},
+  volume={8},
+  number={85},
+  pages={5278},
   year={2023}
 }
 
