### 2023-06-08 b23dde7
Ignore mesh read
