### 2023-06-08 f29dd32
Swap from h5 to xdmf checkpointing with write_checkpoint. Add test on restarting
