+++ tests/test_NSfracStepMove.py
@@ -30,24 +30,6 @@ def test_MovingVortex(num_processors):
     assert eval(err[0]) < 5e-3
 
 
-@pytest.mark.parametrize("num_processors_initial, num_processors_restart", [(1, 2), (2, 2)])
-def test_restart_MovingVortex(num_processors_initial, num_processors_restart):
-    cmd = ("mpirun -np {} oasism NSfracStepMove problem=MovingVortex T=1 Nx=20 Ny=20 checkpoint=20;" +
-           "mpirun -np {} oasism NSfracStepMove problem=MovingVortex T=2 Nx=20 Ny=20 " +
-           "restart_folder=results_moving_vortex/data/1/Checkpoint"
-           )
-    d = subprocess.check_output(cmd.format(num_processors_initial, num_processors_restart), shell=True)
-    match = re.findall("u0=" + number, str(d))
-
-    # Check errors from first and second (restart) simulation
-    tol_initial = 5E-3
-    tol_restart = 5E-4
-
-    assert eval(match[0]) < tol_initial
-    assert eval(match[-1]) < tol_restart
-
-
 if __name__ == '__main__':
     test_MovingVortex(1)
     test_DrivenCavity_with_NSfracStepMove(1)
-    test_restart_MovingVortex(1, 1)
