+++ .github/workflows/check_and_test_package.yml
@@ -64,19 +64,19 @@ jobs:
           python3 -m pip install pip setuptools --upgrade
 
       # See: https://github.com/marketplace/actions/setup-miniconda
-      - name: Setup miniconda
+      - name: Setup Mambaforge
         uses: conda-incubator/setup-miniconda@v2
         with:
-          auto-update-conda: true
           miniforge-variant: Mambaforge
-          channels: conda-forge
+          miniforge-version: latest
           python-version: 3.8
+          channels: conda-forge
           activate-environment: oasismove
-          environment-file: environment.yml
           use-mamba: true
 
       - name: Set cache date
         run: echo "DATE=$(date +'%Y%m%d')" >> $GITHUB_ENV
+
 
       - uses: actions/cache@v2
         with:
