+++ .github/workflows/deploy_pages.yml
@@ -31,17 +31,15 @@ jobs:
       - name: Upgrade pip and setuptools
         run: |
           python3 -m pip install pip setuptools --upgrade
-          
-            # See: https://github.com/marketplace/actions/setup-miniconda
-      - name: Setup miniconda
+
+        # See: https://github.com/marketplace/actions/setup-miniconda
+      - name: Setup Mambaforge
         uses: conda-incubator/setup-miniconda@v2
         with:
-          auto-update-conda: true
           miniforge-variant: Mambaforge
+          miniforge-version: latest
           channels: conda-forge
-          python-version: 3.8
           activate-environment: oasismove
-          environment-file: environment.yml
           use-mamba: true
 
       - name: Set cache date
