+++ tests/test_NSCoupled.py
@@ -6,8 +6,7 @@ number = "[+-]?([0-9]+.[0-9]+e[+-][0-9]+)"
 
 def test_default_Coupled():
     d = subprocess.check_output(
-        "mpirun -np 1 oasism NSCoupled problem=DrivenCavity Nx=25 Ny=25 testing=True",
-                                shell=True
+        "mpirun -np 1 oasism NSCoupled problem=DrivenCavity Nx=25 Ny=25 testing=True", shell=True
     )
     match = re.search("Velocity in corner = " + number, str(d))
     err = match.groups()
@@ -18,8 +17,7 @@ def test_default_Coupled():
 
 def test_default_CR_Coupled():
     d = subprocess.check_output(
-        "mpirun -np 1 oasism NSCoupled problem=DrivenCavity Nx=30 Ny=30 testing=True element=CR",
-                                shell=True
+        "mpirun -np 1 oasism NSCoupled problem=DrivenCavity Nx=30 Ny=30 testing=True element=CR", shell=True
     )
     match = re.search("Velocity in corner = " + number, str(d))
     err = match.groups()
