### 2023-06-09 3879235
Store boundary during checkpointing
