### 2023-06-09 80a79b9
Remove print statement
