### 2023-06-09 a3050cb
Add numpy import with alias
