+++ src/oasismove/NSfracStepMove.py
@@ -33,6 +33,8 @@ import importlib
 import pickle
 from pprint import pprint
 
+import numpy as np
+
 from oasismove.common import *
 
 commandline_kwargs = parse_command_line()
@@ -204,7 +206,7 @@ while t < (T - tstep * DOLFIN_EPS) and not stop:
     t += dt
     tstep += 1
     inner_iter = 0
-    udiff = array([1e8])  # Norm of velocity change over last inner iter
+    udiff = np.array([1e8])  # Norm of velocity change over last inner iter
     num_iter = max(iters_on_first_timestep, max_iter) if tstep <= max_tstep else max_iter
 
     start_timestep_hook(**vars())
