+++ .github/workflows/deploy_pages.yml
@@ -69,7 +69,8 @@ jobs:
 
       # Deploy the book's HTML to gh-pages branch
       - name: Deploy documentation
-          uses: peaceiris/actions-gh-pages@v3
-          if: github.ref == 'refs/heads/main'
+        uses: peaceiris/actions-gh-pages@v3
+        if: github.ref == 'refs/heads/main'
+        with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
           publish_dir: ./docs/_build/html
