+++ docs/stenosis.md
@@ -120,7 +120,9 @@ The velocity field for the stenosis problem without (top) and with (bottom) back
 arrows that have been scaled by the velocity magnitude. 
 ```
 
-```{bibliography}
+## References
+
+```{bibliography} references.bib
+:style: unsrt
 :filter: docname in docnames
 ```
-
