+++ docs/movingtg.md
@@ -101,6 +101,5 @@ $ oasism NSfracStepMove problem=MovingTaylorGreen3D Nx=50 Ny=50 Nz=50
 ## References
 
 ```{bibliography} references.bib
-:style: unsrt
 :filter: docname in docnames
 ```
