+++ _sources/getting_started.md
@@ -37,6 +37,8 @@ When loading a problem file, OasisMove will start by looking inside the `problem
 look in the current working directory. If the problem you have specified after `problem=` is not located in either locations, an error will occur.
 ```
 
-```{bibliography}
+## References
+
+```{bibliography} references.bib
 :filter: docname in docnames
 ```
