### 2023-06-28 #5506 @Kevin-Mattheus-Moerman
:wave: @albansouche, @piyueh, @TormodLandet, would any of you be willing to review this submission for [The Journal of Open Source Softwqare (JOSS)](https://joss.theoj.org/about)? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html.

The review focuses on [the software(https://github.com/KVSlab/OasisMove), and a [short paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05506/joss.05506/10.21105.joss.05506.pdf).

If you are interested please let me by responding in this pre-preview issue. Thanks!
