+++ src/oasismove/problems/NSfracStep/MovingVentricle.py
@@ -1,0 +1,335 @@
+import os
+import pickle
+
+from scipy.interpolate import splrep, splev
+from scipy.spatial import distance
+
+from oasismove.problems.NSfracStep import *
+from oasismove.problems.NSfracStep.MovingCommon import get_visualization_writers
+
+
+def problem_parameters(commandline_kwargs, NS_parameters, **NS_namespace):
+    """
+    Problem file for running CFD simulation in an idealized 3D left ventricle, based on the model presented by
+    Vedula et al.[1] A plug velocity profile is applied at the inlet (mitral valve), and the outlet (aortic valve)
+    is modelled as open wherea pressure condition is applied. Flow rates for the inlet condition is located in the
+    file LV_flowrate.txt. The simulation is run for two cycles (adjustable), where the cardiac cycle is set to 0.9.
+
+    [1] Vedula, V., Fortini, S., Seo, J. H., Querzoli, G., & Mittal, R. (2014). Computational modeling and validation
+    of intraventricular flow in a simple model of the left ventricle. Theoretical and Computational Fluid
+    Dynamics, 28, 589-604.
+    """
+    if "restart_folder" in commandline_kwargs.keys():
+        restart_folder = commandline_kwargs["restart_folder"]
+        restart_folder = path.join(os.getcwd(), restart_folder)
+        f = open(path.join(restart_folder, 'params.dat'), 'rb')
+        NS_parameters.update(pickle.load(f))
+        NS_parameters['restart_folder'] = restart_folder
+    else:
+        # Override some problem specific parameters
+        Re = 3475.59
+        d_mitral = 0.42857
+        u_mean = 15.516
+        nu = 0.00191
+        cardiac_cycle = 0.9
+        number_of_cycles = 1
+        Re_computed = u_mean * d_mitral / nu
+        if MPI.rank(MPI.comm_world) == 0:
+            print("Real RE={} | Computed RE={}".format(Re, Re_computed))
+        NS_parameters.update(
+            # Enable backflow stabilization
+            backflow_facets=[1],
+            backflow_beta=1.0,
+            # Mesh parameters
+            id_wall=0,
+            id_aorta=1,
+            id_mitral=2,
+            d_mitral=d_mitral,
+            # Simulation parameters
+            u_mean=u_mean,
+            number_of_cycles=number_of_cycles,
+            cardiac_cycle=cardiac_cycle,  # s
+            T=cardiac_cycle * number_of_cycles,
+            dt=0.9 / 500,
+            nu=nu,  # cm^2/s
+            dynamic_mesh=True,
+            checkpoint=500,
+            save_step_problem=2,
+            save_step_problem_h5=20E10,
+            save_flow_metrics_tstep=40E10,
+            folder="results_moving_ventricle",
+            mesh_path="src/oasismove/mesh/MovingVentricle3D/Ventricle3D/wall_mesh_003.xml.gz",
+            velocity_degree=1,
+            pressure_degree=1,
+            print_intermediate_info=20,
+            max_iter=2,
+            use_krylov_solvers=True)
+
+
+def mesh(mesh_path, u_mean, dt, velocity_degree, **NS_namespace):
+    # Read mesh and print mesh information
+    mesh = Mesh(mesh_path)
+    print_mesh_information(mesh)
+
+    # Compute mesh information and CFL
[…]
