### 2023-07-02 ab065d8
Increase NSCoupled test resolution
