+++ src/oasismove/problems/NSfracStep/MovingAtrium.py
@@ -1,0 +1,436 @@
+import json
+import pickle
+from os import makedirs
+
+from scipy.interpolate import splrep, splev
+from scipy.spatial import KDTree
+from vampy.simulation.Womersley import make_womersley_bcs, compute_boundary_geometry_acrn
+from vampy.simulation.simulation_common import store_u_mean, get_file_paths, print_mesh_information
+
+from oasismove.problems.NSfracStep import *
+from oasismove.problems.NSfracStep.MovingCommon import get_visualization_writers
+from oasismove.problems.NSfracStep.Probe import Probes  # type: ignore
+
+
+# Override some problem specific parameters
+def problem_parameters(commandline_kwargs, NS_parameters, **NS_namespace):
+    """
+    Problem file for running CFD simulation in left atrial models consisting of arbitrary number of pulmonary veins (PV)
+    (normally 3 to 7), and one outlet being the mitral valve. A Womersley velocity profile is applied at the inlets,
+    where the total flow rate is split between the area ratio of the PVs. The mitral valve is considered open with a
+    constant pressure of p=0. Flow rate and flow split values for the inlet condition are computed from the
+    pre-processing script automatedPreProcessing.py. The simulation is run for two cycles (adjustable), but only the
+    results/solutions from the second cycle are stored to avoid non-physiological effects from the first cycle.
+    One cardiac cycle is set to 0.951 s from [1], and scaled by a factor of 1000, hence all parameters are in
+    [mm] or [ms].
+    The script has been adjusted to handle moving domains.
+
+    [1] Hoi, Yiemeng, et al. "Characterization of volumetric flow rate waveforms at the carotid bifurcations of older
+        adults." Physiological measurement 31.3 (2010): 291.
+    """
+
+    if "restart_folder" in commandline_kwargs.keys():
+        restart_folder = commandline_kwargs["restart_folder"]
+        mesh_path = commandline_kwargs["mesh_path"]
+        f = open(path.join(restart_folder, 'params.dat'), 'rb')
+        NS_parameters.update(pickle.load(f))
+        NS_parameters['restart_folder'] = restart_folder
+        NS_parameters['mesh_path'] = mesh_path
+        globals().update(NS_parameters)
+    else:
+        # Override some problem specific parameters
+        # Parameters are in mm and ms
+        cardiac_cycle = float(commandline_kwargs.get("cardiac_cycle", 1000))
+        number_of_cycles = float(commandline_kwargs.get("number_of_cycles", 1))
+
+        NS_parameters.update(
+            # Moving atrium parameters
+            dynamic_mesh=False,  # Run moving mesh simulation
+            compute_velocity_and_pressure=True,  # Only solve mesh equations
+            # Backflow parameters
+            backflow_beta=0.2,
+            backflow_facets=[],
+            # Fluid parameters
+            nu=3.3018868e-3,  # Viscosity [nu: 0.0035 Pa-s / 1060 kg/m^3 = 3.3018868E-6 m^2/s == 3.3018868E-3 mm^2/ms]
+            # Geometry parameters
+            id_in=[],  # Inlet boundary ID
+            id_out=[],  # Outlet boundary IDs
+            # Simulation parameters
+            cardiac_cycle=cardiac_cycle,  # Run simulation for 1 cardiac cycles [ms]
+            # FIXME: For scaling only
[…]
