### 2023-07-07 b9ad3d8
Add probe to __init__
