### 2023-07-07 6fba67d
Add motion to MovingAtrium
