+++ src/oasismove/NSfracStepMove.py
@@ -346,5 +346,9 @@ theend_hook(**vars())
 
 # Run times
 RunTimePath = os.path.join(newfolder, f"RunTime_{problemname}.csv")
-data = np.array([Time, TentativeVelocity, MeshEquations, PressureSolve, VelocityUpdate])
-np.savetxt(RunTimePath, data.T, delimiter=",")
+rank = MPI.rank(MPI.comm_world)
+if rank == 0:
+    for arr in [Time, TentativeVelocity, MeshEquations, PressureSolve, VelocityUpdate]:
+        print(f"RANK={rank} LEN={len(arr)}")
+    data = np.array([Time, TentativeVelocity, MeshEquations, PressureSolve, VelocityUpdate])
+    np.savetxt(RunTimePath, data.T, delimiter=",")
