### 2023-07-12 #5599 @Kevin-Mattheus-Moerman
@hkjeldsberg our editorial board has some concerns about this publication: https://doi.org/10.1002/cnm.3703 and its relation to pursuing a JOSS publication. That publication's core purpose does seem to be to communicate this software package, and hence the board is concerned a JOSS publication would be a duplication and hence not appropriate. A prior software publication can be permissible, but only for a substantially new version (e.g. featuring significant expansions/enhancements/improvements). Given that the above publication is so recent it would appear that the pursued JOSS publication would very much cover the same software.
Can you please formulate a detailed reply on this and clarify the exact purpose of the above paper: https://doi.org/10.1002/cnm.3703, and contrast it with the purpose of the intended JOSS paper? We appreciate the former may be a lot more comprehensive and features a detailed validation/verification, but what aspects in the JOSS paper are not covered by this published work? We also appreciate that JOSS publications feature specific software review, whereas the journal of the published work may not have featured that, however, this in its own right does not warrant a JOSS publication, if the published work's purpose was to present a software achievement, then that is still a duplicate of a potential JOSS publication.
I am afraid that, if there is too much overlap with the published work, then this submission will need to be retracted/rejected from JOSS. I will now pause this review until the above is addressed and the editorial board has had a chance to review your feedback. Thanks.
