### 2023-08-11 #5599 @hkjeldsberg
Hi @Kevin-Mattheus-Moerman ,

Thank you for your detailed feedback regarding the concerns raised about our recent publication https://doi.org/10.1002/cnm.3703 and our submission to JOSS.

To clarify, while both publications do revolve around the same software package, their intent and focus are considerably different. The aforementioned publication in IJNMBE https://doi.org/10.1002/cnm.3703 primarily considers the validation, verification, and a comprehensive overview of OasisMove’s core functionalities. It provides detailed insight into the mathematical foundation and the technicalities that work as the building blocks of OasisMove. The aim was to showcase its robustness, reliability, and the scientific qualities of our approach.

In contrast, the JOSS submission we have prepared is tailored towards a more concise representation of the software. Our intention was to provide the community with a guide to OasisMove’s practical usability, its broader applications, and to highlight aspects that make it distinct and valuable to potential users. Additionally, as you pointed out, the JOSS publication involves a specific software review which our software hasn’t undergone in the previous publication.

However, considering the concerns of the editorial board, we recognize the points raised regarding potential overlap. Therefore, we agree that the works may appear too similar. We are open to a redaction of our current submission, and if deemed suitable, a potential resubmission in the future after ensuring there is a clear distinction between the two works.

Please let us know the best way to proceed, and thank you for your patience and understanding.

Best regards,
Henrik Kjeldsberg
