+++ tests/test_NSfracStep.py
@@ -41,7 +41,7 @@ def test_spatial_rate_of_convergence(num_processors, solver):
 @pytest.mark.parametrize("solver", ["IPCS_ABCN", "IPCS_ABE", "Chorin", "BDFPC_Fast"])
 @pytest.mark.parametrize("num_processors", [1, 2])
 def test_TaylorGreen2D(num_processors, solver):
-    cmd = ("mpirun -np {} oasism NSfracStep solver={} "
+    cmd = ("mpirun -np {} oasismove NSfracStep solver={} "
            "problem=TaylorGreen2D T=0.01 Nx=30 Ny=30")
     if num_processors > 1 and solver == "Chorin":  # Uses direct solver
         return
@@ -81,7 +81,7 @@ def test_TaylorGreen2D(num_processors, solver):
 
 @pytest.mark.parametrize("num_processors", [1, 2])
 def test_DrivenCavity(num_processors):
-    cmd = ("mpirun -np {} oasism NSfracStep problem=DrivenCavity T=0.01 " +
+    cmd = ("mpirun -np {} oasismove NSfracStep problem=DrivenCavity T=0.01 " +
            "Nx=10 Ny=10 plot_interval=10000 solver={} testing=True")
     d = subprocess.check_output(cmd.format(num_processors, "IPCS_ABCN"), shell=True)
     match = re.search("Velocity norm = " + number, str(d))
