+++ src/oasismove/NSfracStepMove.py
@@ -208,12 +208,12 @@ while t < (T - tstep * DOLFIN_EPS) and not stop:
     inner_iter = 0
     udiff = np.array([1e8])  # Norm of velocity change over last inner iter
     num_iter = max(iters_on_first_timestep, max_iter) if tstep <= max_tstep else max_iter
-    break
 
     start_timestep_hook(**vars())
 
     update_boundary_conditions(**vars())
 
+    break
     # Solve for mesh velocity and move mesh with prescribed motion
     if dynamic_mesh:
         t0 = OasisTimer("Mesh equations")
