+++ src/oasismove/NSfracStepMove.py
@@ -289,13 +289,12 @@ while t < (T - tstep * DOLFIN_EPS) and not stop:
             x_1[ci].axpy(1., x_[ci])
 
     # Print some information
-    if tstep % print_intermediate_info == 0:
-        toc = tx.stop()
-        info_green('Time = {0:2.4e}, timestep = {1:6d}, End time = {2:2.4e}'.format(t, tstep, T))
-        info_red('Total computing time on previous {0:d} timesteps = {1:f}'.format(
-            print_intermediate_info, toc))
-        list_timings(TimingClear.clear, [TimingType.wall])
-        tx.start()
+    toc = tx.stop()
+    info_green('Time = {0:2.4e}, timestep = {1:6d}, End time = {2:2.4e}'.format(t, tstep, T))
+    info_red('Total computing time on previous {0:d} timesteps = {1:f}'.format(
+        print_intermediate_info, toc))
+    list_timings(TimingClear.clear, [TimingType.wall])
+    tx.start()
 
     # AB projection for pressure on next timestep
     if AB_projection_pressure and t < (T - tstep * DOLFIN_EPS) and not stop:
