+++ src/oasismove/NSfracStepMove.py
@@ -301,6 +301,7 @@ while t < (T - tstep * DOLFIN_EPS) and not stop:
     # AB projection for pressure on next timestep
     if AB_projection_pressure and t < (T - tstep * DOLFIN_EPS) and not stop:
         x_['p'].axpy(0.5, dp_.vector())
+    print("Done at ", t)
 exit()
 total_timer.stop()
 list_timings(TimingClear.keep, [TimingType.wall])
