+++ src/oasismove/problems/NSfracStep/DrivenCavity.py
@@ -65,15 +65,7 @@ def pre_solve_hook(mesh, newfolder, velocity_degree, **NS_namespace):
 
 
 def temporal_hook(viz_u, viz_p, tstep, u_, t, uv, p_, plot_interval, testing, **NS_namespace):
-    if tstep % plot_interval == 0 and not testing:
-        print("Writing to file")
-        assign(uv.sub(0), u_[0])
-        assign(uv.sub(1), u_[1])
-        print("Subbed")
-        viz_u.write(uv, t)
-        viz_p.write(p_, t)
-        print("Completed")
-
+    pass
 
 def theend_hook(u_, uv, mesh, testing, **NS_namespace):
     u_norm = norm(u_[0].vector())
