+++ src/oasismove/problems/NSfracStep/DrivenCavity.py
@@ -71,7 +71,7 @@ def temporal_hook(viz_u, viz_p, tstep, u_, t, uv, p_, plot_interval, testing, **
         assign(uv.sub(0), u_[0])
         assign(uv.sub(1), u_[1])
         print("Subbed")
-        #viz_u.write(uv, t)
+        viz_u.write(uv, t)
         #viz_p.write(p_, t)
         print("Completed")
 
