+++ src/oasismove/problems/NSfracStep/DrivenCavity.py
@@ -82,9 +82,10 @@ def temporal_hook(viz_u, viz_p, newfolder, tstep, u_, t, uv, p_, plot_interval, 
         assign(uv.sub(1), u_[1])
 
         print("Writing")
-        viz_u = HDF5File(MPI.comm_world, u_path, file_mode=file_mode)
-        viz_u.write(uv, "/velocity", tstep)
-        viz_u.close()
+        if MPI.rank(MPI.comm_world) == 0:
+            viz_u = HDF5File(MPI.comm_world, u_path, file_mode=file_mode)
+            viz_u.write(uv, "/velocity", tstep)
+            viz_u.close()
         out_file = path.join(newfolder, "Solutions", "out.txt")
         # with open(out_file, 'a') as out:
         #     out.write(str(tstep) + "\n")
