+++ src/oasismove/problems/NSfracStep/DrivenCavity.py
@@ -31,7 +31,7 @@ def problem_parameters(NS_parameters, **NS_namespace):
 
 
 # Create a mesh
-def mesh(Nx=50, Ny=50, newfolder, **params):
+def mesh(newfolder,Nx=50, Ny=50,  **params):
     m = UnitSquareMesh(Nx, Ny)
     file_mode = "w"
     u_path = path.join(newfolder, "Solutions", "u.h5")
