+++ src/oasismove/problems/NSfracStep/DrivenCavity.py
@@ -33,10 +33,6 @@ def problem_parameters(NS_parameters, **NS_namespace):
 # Create a mesh
 def mesh(Nx=50, Ny=50,  **params):
     m = UnitSquareMesh(Nx, Ny)
-    file_mode = "w"
-    u_path = "u.h5"
-    viz_u = HDF5File(MPI.comm_world, u_path, file_mode=file_mode)
-
     return m
 
 
