+++ src/oasismove/problems/NSfracStep/DrivenCavity.py
@@ -6,6 +6,7 @@ from oasismove.problems.NSfracStep import *
 from oasismove.problems.NSfracStep.MovingCommon import get_visualization_writers
 
 from dolfin import MPI
+comm = MPI.comm_world
 
 
 # Override some problem specific parameters
@@ -84,10 +85,11 @@ def temporal_hook(viz_u, viz_p, newfolder, tstep, u_, t, uv, p_, plot_interval, 
         assign(uv.sub(1), u_[1])
 
         print("Writing")
-        if MPI.rank(MPI.comm_world) == 0:
-            file_mode = "w"
-            u_path = path.join(newfolder, "Solutions", "u.h5")
-            viz_u = HDF5File(MPI.comm_world, u_path, file_mode=file_mode)
+        file_mode = "w"
+        u_path = path.join(newfolder, "Solutions", "u.h5")
+
+        MPI.barrier(comm)
+        viz_u = HDF5File(MPI.comm_world, u_path, file_mode=file_mode)
             # viz_u.write(uv, "/velocity", tstep)
             # viz_u.close()
         out_file = path.join(newfolder, "Solutions", "out.txt")
