+++ src/oasismove/problems/NSfracStep/DrivenCavity.py
@@ -69,11 +69,6 @@ def pre_solve_hook(mesh, newfolder, velocity_degree, **NS_namespace):
 
 
 def temporal_hook(viz_u, viz_p, newfolder, tstep, u_, t, uv, p_, plot_interval, testing, **NS_namespace):
-    common_path = path.join(newfolder, "Solutions")
-
-    if MPI.rank(MPI.comm_world) == 0:
-        if not path.exists(common_path):
-            makedirs(common_path)
 
     if tstep % 5 == 0:
         print("Writing to file")
@@ -90,11 +85,9 @@ def temporal_hook(viz_u, viz_p, newfolder, tstep, u_, t, uv, p_, plot_interval, 
 
         MPI.barrier(comm)
         viz_u = HDF5File(MPI.comm_world, u_path, file_mode=file_mode)
-            # viz_u.write(uv, "/velocity", tstep)
-            # viz_u.close()
-        out_file = path.join(newfolder, "Solutions", "out.txt")
-        # with open(out_file, 'a') as out:
-        #     out.write(str(tstep) + "\n")
+        viz_u.write(uv, "/velocity", tstep)
+        viz_u.close()
+        MPI.barrier(comm)
         print("Completed")
 
 
