+++ src/oasismove/problems/NSfracStep/DrivenCavity.py
@@ -5,7 +5,6 @@ from os import makedirs
 from oasismove.problems.NSfracStep import *
 from oasismove.problems.NSfracStep.MovingCommon import get_visualization_writers
 
-from dolfin import MPI
 comm = MPI.comm_world
 
 
@@ -83,10 +82,11 @@ def temporal_hook(viz_u, viz_p, newfolder, tstep, u_, t, uv, p_, plot_interval, 
         file_mode = "w"
         u_path = path.join(newfolder, "Solutions", "u.h5")
 
+        MPI.barrier(comm)
         viz_u = HDF5File(MPI.comm_world, u_path, file_mode=file_mode)
+        MPI.barrier(comm)
         # viz_u.write(uv, "/velocity", tstep)
         # viz_u.close()
-        MPI.barrier(comm)
         print("Completed")
 
 
