+++ src/oasismove/problems/NSfracStep/DrivenCavity.py
@@ -83,7 +83,10 @@ def temporal_hook(viz_u, viz_p, newfolder, tstep, u_, t, uv, p_, plot_interval, 
         u_path = path.join(newfolder, "Solutions", "u.h5")
 
         MPI.barrier(comm)
-        viz_u = HDF5File(MPI.comm_world, u_path, file_mode=file_mode)
+        quantity ="velocity"
+        viz = XDMFFile(MPI.comm_world, path.join(newfolder, "Solutions", f"{quantity}.xdmf"))
+
+        #viz_u = HDF5File(MPI.comm_world, u_path, file_mode=file_mode)
         MPI.barrier(comm)
         # viz_u.write(uv, "/velocity", tstep)
         # viz_u.close()
