+++ mwe.py
@@ -1,4 +1,4 @@
-from dolfin import MPI, HDF5File
+from dolfin import MPI, XDMFFile
 
 comm = MPI.comm_world
 u_path = "velocity.h5"
@@ -8,7 +8,9 @@ for i in range(10):
         print("Writing")
 
     MPI.barrier(comm)
-    viz_u = HDF5File(MPI.comm_world, u_path, file_mode=file_mode)
+    # viz_u = HDF5File(MPI.comm_world, u_path, file_mode=file_mode)
+    viz = XDMFFile(MPI.comm_world, "u.xdmf")
+
     MPI.barrier(comm)
 
     if MPI.rank(comm) == 0:
