+++ mwe.py
@@ -1,8 +1,14 @@
-from dolfin import MPI, XDMFFile
+from dolfin import MPI, XDMFFile, UnitSquareMesh, FunctionSpace, Function
 
 comm = MPI.comm_world
 u_path = "velocity.h5"
 file_mode = "w"
+
+# Create function
+mesh = UnitSquareMesh(1, 1)
+V = FunctionSpace(mesh, "CG", 1)
+u = Function(V)
+t = 0.0
 
 if MPI.rank(comm) == 0:
     print("Writing")
@@ -10,7 +16,7 @@ if MPI.rank(comm) == 0:
 MPI.barrier(comm)
 # viz_u = HDF5File(MPI.comm_world, u_path, file_mode=file_mode)
 viz = XDMFFile(MPI.comm_world, "u.xdmf")
-
+viz.write(u, t)
 MPI.barrier(comm)
 
 if MPI.rank(comm) == 0:
