+++ mwe_xdmf.py
@@ -1,4 +1,4 @@
-from dolfin import MPI, XDMFFile, UnitSquareMesh, FunctionSpace, Function, HDF5File
+from dolfin import MPI, XDMFFile, UnitSquareMesh, FunctionSpace, Function
 
 comm = MPI.comm_world
 u_path = "velocity.xdmf"
@@ -9,10 +9,6 @@ V = FunctionSpace(mesh, "CG", 1)
 u = Function(V)
 t = 0.0
 
-# Code hangs here:
-
 viz = XDMFFile(comm, u_path)
-viz.write(u, t)  # Hangs
-
-if MPI.rank(comm) == 0:
-    print("Done")
+# Code hangs here:
+viz.write(u, t)
