### 2023-09-11 #5599 @Kevin-Mattheus-Moerman
@hkjeldsberg thanks for clarifying the nature of your prior publication. I have consulted the editorial board on this. Unfortunately we do believe that, as it stands, the work submitted here is deemed too similar to that reviewed/presented in that prior paper (https://doi.org/10.1002/cnm.3703).

The prior publication does appear to be a "software publication" in that the core purpose was to present the software and capabilities (in detail). Although the nature of JOSS papers/reviews is different, and the level of detail presented in the prior paper is much higher, it does represent a publication of the software. A JOSS publication on top of the prior paper would be justified if the software were to contain sufficient additional/novel functionality on top of the previously published version. Since this is not the case we feel a JOSS publication at this point is not justified.

We will therefore now proceed to reject this paper.

Note we do welcome a re-submission once your software has more significantly departed from the state at the prior publication e.g. in terms of functionality/capabilities/performance.
